Move FilterControl back to ResourceList
