+++ src/components/ResourceList/components/index.ts
@@ -1,3 +1,5 @@
 export * from './BulkActions';
 
 export * from './CheckableButton';
+
+export * from './FilterControl';
